requirements.txt
procgen>=0.10.7
tqdm==4.66.5
pytorch-lightning==2.0.1
opencv-python==4.10.0.84
transformers==4.19.1
triton==2.0.0
einops==0.8.0
piq==0.8.0
xformers>=0.0.20
lightning==2.0.8
kmeans-pytorch==0.3
jsonargparse[signatures]>=4.17.0
tensorboard==2.17.1
tensorboardX==2.6.2.2
deepspeed==0.10.1
natsort>=8.4.0
omegaconf>=2.3.0
safetensors==0.5.1
kornia==0.7.4
open-clip-torch>=2.20.0
scipy==1.14.1
scikit-learn==1.6.1
setuptools<70.0.0
matplotlib==3.10.1
